FINAL COMMIT DAY 5
-Added possibility for user to customize the amount of columns and rows in the scene
-Added user input validation
-edited README file

+++ src/util/Tile.java
@@ -7,12 +7,12 @@ public class Tile {
     
     public Point2D.Float topLeft ;
     
-    public int width;
-    public int height;
+    public float width;
+    public float height;
     
     public String type;
     
-    public Tile(String t,int X, int Y, int w, int h){
+    public Tile(String t,float X, float Y, float w, float h){
         type = t;
         topLeft = new Point2D.Float(X,Y);
         width = w;
